Refactor names a bit
Got some good feedback from @quartzmo on these class names.
