+++ google-cloud-speech/test/google/cloud/speech/result_test.rb
@@ -14,10 +14,10 @@
 
 require "helper"
 
-describe Google::Cloud::Speech::Results, :mock_speech do
+describe Google::Cloud::Speech::Result, :mock_speech do
   let(:results_json) { "{\"results\":[{\"alternatives\":[{\"transcript\":\"how old is the Brooklyn Bridge\",\"confidence\":0.98267895}]}]}" }
   let(:results_grpc) { Google::Cloud::Speech::V1beta1::SyncRecognizeResponse.decode_json results_json }
-  let(:results) { Google::Cloud::Speech::Results.from_grpc results_grpc }
+  let(:results) { results_grpc.results.map { |result_grpc| Google::Cloud::Speech::Result.from_grpc result_grpc } }
 
   it "knows itself" do
     results.count.must_equal 1
